+++ src/random_ray/random_ray.cpp
@@ -20,28 +20,28 @@ namespace openmc {
 // returns 1 - exp(-tau)
 // Equivalent to -(_expm1f(-tau)), but faster
 // Written by Colin Josey.
-float cjosey_exponential(float tau)
+double cjosey_exponential(double tau)
 {
-  constexpr float c1n = -1.0000013559236386308f;
-  constexpr float c2n = 0.23151368626911062025f;
-  constexpr float c3n = -0.061481916409314966140f;
-  constexpr float c4n = 0.0098619906458127653020f;
-  constexpr float c5n = -0.0012629460503540849940f;
-  constexpr float c6n = 0.00010360973791574984608f;
-  constexpr float c7n = -0.000013276571933735820960f;
+  constexpr double c1n = -1.0000013559236386308;
+  constexpr double c2n = 0.23151368626911062025;
+  constexpr double c3n = -0.061481916409314966140;
+  constexpr double c4n = 0.0098619906458127653020;
+  constexpr double c5n = -0.0012629460503540849940;
+  constexpr double c6n = 0.00010360973791574984608;
+  constexpr double c7n = -0.000013276571933735820960;
 
-  constexpr float c0d = 1.0f;
-  constexpr float c1d = -0.73151337729389001396f;
-  constexpr float c2d = 0.26058381273536471371f;
-  constexpr float c3d = -0.059892419041316836940f;
-  constexpr float c4d = 0.0099070188241094279067f;
-  constexpr float c5d = -0.0012623388962473160860f;
-  constexpr float c6d = 0.00010361277635498731388f;
-  constexpr float c7d = -0.000013276569500666698498f;
+  constexpr double c0d = 1.0;
+  constexpr double c1d = -0.73151337729389001396;
+  constexpr double c2d = 0.26058381273536471371;
+  constexpr double c3d = -0.059892419041316836940;
+  constexpr double c4d = 0.0099070188241094279067;
+  constexpr double c5d = -0.0012623388962473160860;
+  constexpr double c6d = 0.00010361277635498731388;
+  constexpr double c7d = -0.000013276569500666698498;
 
-  float x = -tau;
+  double x = -tau;
 
-  float den = c7d;
+  double den = c7d;
   den = den * x + c6d;
   den = den * x + c5d;
   den = den * x + c4d;
@@ -50,7 +50,7 @@ float cjosey_exponential(float tau)
   den = den * x + c1d;
   den = den * x + c0d;
 
-  float num = c7n;
+  double num = c7n;
   num = num * x + c6n;
   num = num * x + c5n;
   num = num * x + c4n;
@@ -92,37 +92,37 @@ float cjosey_exponential(float tau)
 // approximation. It is accurate to 2e-7 over [0, 1e5]. Developed by Colin
 // Josey using Remez's algorithm, with original implementation in OpenMOC at:
 // https://github.com/mit-crpg/OpenMOC/blob/develop/src/exponentials.h
-float exponentialG(float tau)
+double exponentialG(double tau)
 {
   // Numerator coefficients in rational approximation for 1/x - (1 - exp(-x)) /
   // x^2
-  constexpr float d0n = 0.5f;
-  constexpr float d1n = 0.176558112351595f;
-  constexpr float d2n = 0.04041584305811143f;
-  constexpr float d3n = 0.006178333902037397f;
-  constexpr float d4n = 0.0006429894635552992f;
-  constexpr float d5n = 0.00006064409107557148f;
+  constexpr double d0n = 0.5;
+  constexpr double d1n = 0.176558112351595;
+  constexpr double d2n = 0.04041584305811143;
+  constexpr double d3n = 0.006178333902037397;
+  constexpr double d4n = 0.0006429894635552992;
+  constexpr double d5n = 0.00006064409107557148;
 
   // Denominator coefficients in rational approximation for 1/x - (1 - exp(-x))
   // / x^2
-  constexpr float d0d = 1.0f;
-  constexpr float d1d = 0.6864462055546078f;
-  constexpr float d2d = 0.2263358514260129f;
-  constexpr float d3d = 0.04721469893686252f;
-  constexpr float d4d = 0.006883236664917246f;
-  constexpr float d5d = 0.0007036272419147752f;
-  constexpr float d6d = 0.00006064409107557148f;
+  constexpr double d0d = 1.0;
+  constexpr double d1d = 0.6864462055546078;
+  constexpr double d2d = 0.2263358514260129;
+  constexpr double d3d = 0.04721469893686252;
+  constexpr double d4d = 0.006883236664917246;
+  constexpr double d5d = 0.0007036272419147752;
+  constexpr double d6d = 0.00006064409107557148;
 
-  float x = tau;
+  double x = tau;
 
-  float num = d5n;
+  double num = d5n;
   num = num * x + d4n;
   num = num * x + d3n;
   num = num * x + d2n;
   num = num * x + d1n;
   num = num * x + d0n;
 
-  float den = d6d;
+  double den = d6d;
   den = den * x + d5d;
   den = den * x + d4d;
   den = den * x + d3d;
@@ -138,38 +138,38 @@ float exponentialG(float tau)
 // [0, 1e6]. Developed by Colin Josey using Remez's algorithm, with original
 // implementation in OpenMOC at:
 // https://github.com/mit-crpg/OpenMOC/blob/develop/src/exponentials.h
-float exponentialG2(float tau)
+double exponentialG2(double tau)
 {
 
   // Coefficients for numerator in rational approximation
-  constexpr float g1n = -0.08335775885589858f;
-  constexpr float g2n = -0.003603942303847604f;
-  constexpr float g3n = 0.0037673183263550827f;
-  constexpr float g4n = 0.00001124183494990467f;
-  constexpr float g5n = 0.00016837426505799449f;
+  constexpr double g1n = -0.08335775885589858;
+  constexpr double g2n = -0.003603942303847604;
+  constexpr double g3n = 0.0037673183263550827;
+  constexpr double g4n = 0.00001124183494990467;
+  constexpr double g5n = 0.00016837426505799449;
 
   // Coefficients for denominator in rational approximation
-  constexpr float g1d = 0.7454048371823628f;
-  constexpr float g2d = 0.23794300531408347f;
-  constexpr float g3d = 0.05367250964303789f;
-  constexpr float g4d = 0.006125197988351906f;
-  constexpr float g5d = 0.0010102514456857377f;
+  constexpr double g1d = 0.7454048371823628;
+  constexpr double g2d = 0.23794300531408347;
+  constexpr double g3d = 0.05367250964303789;
+  constexpr double g4d = 0.006125197988351906;
+  constexpr double g5d = 0.0010102514456857377;
 
-  float x = tau;
+  double x = tau;
 
-  float num = g5n;
+  double num = g5n;
   num = num * x + g4n;
   num = num * x + g3n;
   num = num * x + g2n;
   num = num * x + g1n;
   num = num * x;
 
-  float den = g5d;
+  double den = g5d;
   den = den * x + g4d;
   den = den * x + g3d;
   den = den * x + g2d;
   den = den * x + g1d;
-  den = den * x + 1.0f;
+  den = den * x + 1.0;
 
   return num / den;
 }
@@ -325,19 +325,19 @@ void RandomRay::attenuate_flux_flat_source(double distance, bool is_active, bool
 
   // MOC incoming flux attenuation + source contribution/attenuation equation
   for (int g = 0; g < negroups_; g++) {
-    float sigma_t = domain_->sigma_t_[material * negroups_ + g];
-    float tau = sigma_t * distance;
-    float exponential = cjosey_exponential(tau); // exponential = 1 - exp(-tau)
-    float new_delta_psi =
-      (angular_flux_[g] - (float)domain_->source_regions_.source(sr, g)) * exponential;
+    double sigma_t = domain_->sigma_t_[material * negroups_ + g];
+    double tau = sigma_t * distance;
+    double exponential = cjosey_exponential(tau); // exponential = 1 - exp(-tau)
+    double new_delta_psi =
+      (angular_flux_[g] - domain_->source_regions_.source(sr, g)) * exponential;
     delta_psi_[g] = new_delta_psi;
     angular_flux_[g] -= new_delta_psi;
     if (td_transport) {
-      float sigma_t_td = domain_->sigma_t_td_[material * negroups_ + g];
-      float tau_td = sigma_t_td * distance;
-      float exponential_td = cjosey_exponential(tau_td); // exponential = 1 - exp(-tau)
-      float new_delta_psi_td =
-        (angular_flux_td_[g] - (float)domain_->source_regions_.source_td(sr, g)) * exponential_td;
+      double sigma_t_td = domain_->sigma_t_td_[material * negroups_ + g];
+      double tau_td = sigma_t_td * distance;
+      double exponential_td = cjosey_exponential(tau_td); // exponential = 1 - exp(-tau)
+      double new_delta_psi_td =
+        (angular_flux_td_[g] - domain_->source_regions_.source_td(sr, g)) * exponential_td;
       delta_psi_td_[g] = new_delta_psi_td;
       angular_flux_td_[g] -= new_delta_psi_td;
     }
@@ -423,8 +423,8 @@ void RandomRay::attenuate_flux_linear_source(double distance, bool is_active)
   for (int g = 0; g < negroups_; g++) {
 
     // Compute tau, the optical thickness of the ray segment
-    float sigma_t = domain_->sigma_t_[material * negroups_ + g];
-    float tau = sigma_t * distance;
+    double sigma_t = domain_->sigma_t_[material * negroups_ + g];
+    double tau = sigma_t * distance;
 
     // If tau is very small, set it to zero to avoid numerical issues.
     // The following computations will still work with tau = 0.
@@ -435,21 +435,21 @@ void RandomRay::attenuate_flux_linear_source(double distance, bool is_active)
     // Compute linear source terms, spatial and directional (dir),
     // calculated from the source gradients dot product with local centroid
     // and direction, respectively.
-    float spatial_source =
-      (float)domain_->source_regions_.source(sr, g) +
+    double spatial_source =
+      domain_->source_regions_.source(sr, g) +
       rm_local.dot(domain_->source_regions_.source_gradients(sr, g));
-    float dir_source =
+    double dir_source =
       u().dot(domain_->source_regions_.source_gradients(sr, g));
 
-    float gn = exponentialG(tau);
-    float f1 = 1.0f - tau * gn;
-    float f2 = (2.0f * gn - f1) * distance_2;
-    float new_delta_psi = (angular_flux_[g] - spatial_source) * f1 * distance -
+    double gn = exponentialG(tau);
+    double f1 = 1.0f - tau * gn;
+    double f2 = (2.0f * gn - f1) * distance_2;
+    double new_delta_psi = (angular_flux_[g] - spatial_source) * f1 * distance -
                           0.5 * dir_source * f2;
 
-    float h1 = f1 - gn;
-    float g1 = 0.5f - h1;
-    float g2 = exponentialG2(tau);
+    double h1 = f1 - gn;
+    double g1 = 0.5f - h1;
+    double g2 = exponentialG2(tau);
     g1 = g1 * spatial_source;
     g2 = g2 * dir_source * distance * 0.5f;
     h1 = h1 * angular_flux_[g];
@@ -557,12 +557,12 @@ void RandomRay::initialize_ray(uint64_t ray_id, FlatSourceDomain* domain)
   int64_t sr = domain_->source_region_offsets_[i_cell] + cell_instance();
 
   for (int g = 0; g < negroups_; g++) {
-    angular_flux_[g] = (float)domain_->source_regions_.source(sr, g);
+    angular_flux_[g] = domain_->source_regions_.source(sr, g);
   }
 
   if (settings::run_mode == RunMode::TIME_DEPENDENT) {
     for (int g = 0; g < negroups_; g++) {
-      angular_flux_td_[g] = (float)domain_->source_regions_.source_td(sr, g);
+      angular_flux_td_[g] = domain_->source_regions_.source_td(sr, g);
     }
   }
 }
